pyproject.toml
[project]
name = "paq"
version = "0.0.0"
description = "side project package manager"
authors = [
    {name = "Xavier Mitault", email = "xavier.mitault@epitech.eu"},
]
dependencies = [
    "requests>=2.31.0",
    "xdg-base-dirs>=6.0.1",
    "toml>=0.10.2",
    "rich-argparse>=1.4.0",
    "rich>=13.7.0",
    "PyGithub>=2.2.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.scripts]
build = """
pdm run nuitka3
 --company-name=Saverio976
 --enable-plugins=no-qt
 --enable-plugins=upx
 --product-name=paq
 --follow-imports
 --standalone
 --onefile
 --onefile-tempdir-spec='{CACHE_DIR}/{COMPANY}/{PRODUCT}/unpack'
 src/paq
"""
run = "python ./src/paq"

[tool.pdm.dev-dependencies]
dev = [
    "nuitka>=2.0.2",
    "black>=24.1.1",
    "flake8>=7.0.0",
]

[tool.pyright]
exclude = [ ".venv" ]
venvPath = "."
venv = ".venv"

[tool.black]
line-length = 79
